Move facebook provider to global providers folder
