+++ src/app/auth/module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 
-import { FacebookProvider } from '../app.providers';
+import { FacebookProvider } from '../../providers/facebook';
 
 import { AuthEffects } from './effects/login';
 import { AuthService } from './services/auth';
